requirements.txt
Flask==2.2.3
python-dotenv==1.0.0
tensorflow==2.12.0
Pillow==9.5.0
numpy==1.24.3
boto3==1.26.0
